+++ aptos-move/framework/aptos-framework/sources/account.spec.move
@@ -4,13 +4,6 @@ spec aptos_framework::account {
         pragma aborts_if_is_strict;
     }
 
-    /// Convert address to singer and return.
-    spec create_signer(addr: address): signer {
-        pragma opaque;
-        aborts_if [abstract] false;
-        ensures [abstract] signer::address_of(result) == addr;
-    }
-
     /// Only the address `@aptos_framework` can call.
     /// OriginatingAddress does not exist under `@aptos_framework` before the call.
     spec initialize(aptos_framework: &signer) {
