+++ aptos-move/framework/aptos-framework/sources/genesis.spec.move
@@ -11,11 +11,4 @@ spec aptos_framework::genesis {
     spec initialize_for_verification {
         pragma verify = true;
     }
-
-    spec create_signer(addr: address): signer {
-        use std::signer;
-        pragma opaque;
-        aborts_if false;
-        ensures signer::address_of(result) == addr;
-    }
 }
